Job Cloud/task_a/functions/http_trigger/requirements.txt
google-cloud-storage==2.18.2
